P1/requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
cryptography==36.0.0
idna==3.3
Naked==0.1.31
numpy==1.21.4
pycparser==2.21
pycrypto==2.6.1
PyYAML==6.0
requests==2.26.0
shellescape==3.8.1
snakeviz==2.1.1
tornado==6.1
urllib3==1.26.7
